Restructured the directories to conform with the routes
