+++ v2/src/views/work/WorkExperience.tsx
@@ -1,9 +1,8 @@
-import UppercasedText from "../components/text/UppercasedText";
-import NemesisExperience from "./work/nemesis";
-import Parallaxis2020Experience from "./work/parallaxis2020";
-import Parallaxis2021Experience from "./work/parallaxis2021";
-import SlideIntoViewAnimation from "../components/animations/SlideIntoView";
-import HSSeniorYearResearch from "./research/HSSeniorYearResearch";
+import UppercasedText from "../../components/text/UppercasedText";
+import NemesisExperience from "../work/nemesis";
+import Parallaxis2020Experience from "../work/parallaxis2020";
+import Parallaxis2021Experience from "../work/parallaxis2021";
+import SlideIntoViewAnimation from "../../components/animations/SlideIntoView";
 
 /**
  * Work Experience Component
@@ -24,9 +23,6 @@ function WorkExperience() {
 			</SlideIntoViewAnimation>
 			<SlideIntoViewAnimation>
 				<Parallaxis2020Experience></Parallaxis2020Experience>
-			</SlideIntoViewAnimation>
-			<SlideIntoViewAnimation>
-				<HSSeniorYearResearch></HSSeniorYearResearch>
 			</SlideIntoViewAnimation>
 		</div>
 	);
